% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clean.fnames}
\alias{clean.fnames}
\title{clean.fnames}
\usage{
clean.fnames(fnames)
}
\arguments{
\item{fnames}{feature names}
}
\value{
array of character
}
\description{
clean up feature names to make them legal for modelling. Only keep numbers/letters/_.
}
\examples{
clean.fnames(c("client's birthday","sum assured in thousand (SGD)"))
}
