% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_h2o.R
\name{h2o.save_model}
\alias{h2o.save_model}
\title{h2o.save_model}
\usage{
h2o.save_model(
  model,
  model.dir,
  X.valid.hex,
  id = "LifeID",
  label = "label",
  verbose = TRUE
)
}
\arguments{
\item{model}{h2o.model}

\item{model.dir}{[path] path to save the model}

\item{X.valid.hex}{valid set in h2o.DataFrame}

\item{id}{[string] id column}

\item{label}{[string] label column}

\item{verbose}{whether to show message}
}
\description{
Save h2o.model and extra metrics, related results
}
\examples{
library(h2o)
h2o.init(nthreads = 1,max_mem_size = '1g')
y <- 'am'
x <- setdiff(colnames(mtcars), y)
model.id <- 'h2o.model'
mtcars <- tibble::rownames_to_column(mtcars, 'id')
hex <- as.h2o(mtcars)
hex$am <- h2o.asfactor(hex$am)
h2o.model <- h2o.gbm(x, y , hex, model.id,
                     validation_frame = hex, nfolds = 2,
                     keep_cross_validation_predictions = TRUE,
                     min_rows = 1, max_depth = 3, ntrees = 10)
h2o.save_model(h2o.model, './saved_model', hex, id = 'id', label = 'am')
h2o.shutdown(FALSE)
}
