% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{init.dirs}
\alias{init.dirs}
\title{init.dirs}
\usage{
init.dirs(dirs, verbose = T)
}
\arguments{
\item{dirs}{directory or directories}

\item{verbose}{whether to show messages}
}
\description{
create dirs if not exist
}
\examples{
init.dirs('./saved_model/lgb_model')
}
