% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{json_list}
\alias{json_list}
\alias{list2json}
\alias{json2list}
\title{list2json}
\usage{
list2json(x, path)

json2list(path)
}
\arguments{
\item{x}{list}

\item{path}{path of json file}
}
\description{
write list to json file

read json file as list
}
\examples{
list2json(list(a = 1, b = 2), 'test.json')
json2list('test.json')
}
