% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lightgbm.R
\name{lgb.extract_cv_score}
\alias{lgb.extract_cv_score}
\title{lgb.extract_cv_score}
\usage{
lgb.extract_cv_score(cv, score.fun = Metrics::auc)
}
\arguments{
\item{cv}{\code{lgb.cv} object}

\item{score.fun}{function for scoring, score.fun(label, preds)}
}
\value{
list of (cv.score, cv.preds)
}
\description{
extract cv prediction from lgb.cv object
}
\examples{
library(lightgbm)
data(agaricus.train, package = "lightgbm")
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label = train$label)
params <- list(objective = "regression", metric = "l2")
model <- lgb.cv(
  params = params
  , data = dtrain
  , nrounds = 10L
  , nfold = 3L
  , min_data = 1L
  , learning_rate = 1.0
  , early_stopping_rounds = 5L
)
res <- lgb.extract_cv_score(model, score.fun = Metrics::rmse)
res$cv.score
}
