% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lightgbm.R
\name{lgb.load_model}
\alias{lgb.load_model}
\title{lgb.load_model
Load LGB model}
\usage{
lgb.load_model(model.dir, model.id = "")
}
\arguments{
\item{model.dir}{base path for saving model object and meta-data}

\item{model.id}{identifier for model}
}
\description{
lgb.load_model
Load LGB model
}
