% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lightgbm.R
\name{lgb.predict}
\alias{lgb.predict}
\title{lgb.predict}
\usage{
lgb.predict(bst, df.test, ...)
}
\arguments{
\item{bst}{boosting model from \code{lgb.train.cv}}

\item{df.test}{data.frame of test set}

\item{...}{predcit.lgb.Booster other params}
}
\value{
array of predictions
}
\description{
predict using LGB model
}
\examples{
\dontrun{
preds <- lgb.predict(bst, df.valid)
}

}
