% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lightgbm.R
\name{lgb.save_model}
\alias{lgb.save_model}
\title{lgb.save_model}
\usage{
lgb.save_model(bst, model.dir, model.id = "", verbose = T)
}
\arguments{
\item{bst}{boosting model from \code{lgb.train.cv}}

\item{model.dir}{base path for saving model object and meta-data}

\item{model.id}{identifier for model}

\item{verbose}{whether display saving information, defaults at TRUE}
}
\description{
save trained LGB model together with related meta-data
}
\examples{
\dontrun{
save.lgb.model(bst, './saved_model', 'base_model')
}
}
