% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read.cb}
\alias{read.cb}
\title{read.cb}
\usage{
read.cb(header = T, sep = "\\t")
}
\arguments{
\item{header}{whether table has header}

\item{sep}{separator}
}
\value{
data.frame
}
\description{
read text from clipboard. suitable to copy table from excel to R
}
