% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_h2o.R
\name{save.h2o.cv.pred}
\alias{save.h2o.cv.pred}
\title{save.h2o.cv.pred}
\usage{
save.h2o.cv.pred(model, model.id, model.dir, id, label, verbose = TRUE)
}
\arguments{
\item{model}{h2o.model}

\item{model.id}{[string] model.id}

\item{model.dir}{[path] path to save the model}

\item{id}{[string] id column}

\item{label}{[string] label column}

\item{verbose}{whether to show message}
}
\description{
save.h2o.cv.pred
}
