% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_h2o.R
\name{save.h2o.cv.score}
\alias{save.h2o.cv.score}
\title{save.h2o.cv.score}
\usage{
save.h2o.cv.score(model, model.id, model.dir, verbose = TRUE)
}
\arguments{
\item{model}{h2o.model}

\item{model.id}{[string] model.id}

\item{model.dir}{[path] path to save the model}

\item{verbose}{whether to show message}
}
\description{
save cv score from h2o model
}
