% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_h2o.R
\name{save.h2o.valid.pred}
\alias{save.h2o.valid.pred}
\title{save.h2o.valid.pred}
\usage{
save.h2o.valid.pred(
  model,
  model.id,
  model.dir,
  X.valid.hex,
  id,
  label,
  verbose = TRUE
)
}
\arguments{
\item{model}{h2o.model}

\item{model.id}{[string] model.id}

\item{model.dir}{[path] path to save the model}

\item{X.valid.hex}{valid set in h2o.DataFrame}

\item{id}{[string] id column}

\item{label}{[string] label column}

\item{verbose}{whether to show message}
}
\description{
Save valid set prediction from h2o.model
}
