% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{save.rules}
\alias{save.rules}
\title{save.rules}
\usage{
save.rules(rules, path)
}
\arguments{
\item{rules}{rules to index categorical features}

\item{path}{path for rules json file}
}
\description{
save trained rules for categorical features as json file
}
