% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shap_values_mtcars}
\alias{shap_values_mtcars}
\title{shap value mtcars}
\format{
An object of class \code{data.frame} with 32 rows and 11 columns.
}
\usage{
shap_values_mtcars
}
\description{
shap value matrix of a h2o.GBM model, which takse mpg as label and
the other variables as predictors using mtcars dataset
the code to generate the data is as following:
data(mtcars)
library(h2o)
library(dplyr)
}
\details{
h2o.init()
hex <- as.h2o(mtcars)
model <- h2o.gbm(setdiff(colnames(mtcars), 'mpg'), mpg',min_rows = 3,
training_frame = hex, nfolds = 3)
shap_values_mtcars <- h2o.predict_contributions(model, hex) %>% as.data.frame()
}
\keyword{datasets}
