% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_xgboost.R
\name{xgb.load_model}
\alias{xgb.load_model}
\title{xgb.load_model}
\usage{
xgb.load_model(model.dir, model.id = "")
}
\arguments{
\item{model.dir}{base path for saving model object and meta-data}

\item{model.id}{identifier for model}
}
\description{
load model from saved directory
}
