% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_xgboost.R
\name{xgb.predict}
\alias{xgb.predict}
\title{xgb.predict}
\usage{
xgb.predict(bst, df.test, ...)
}
\arguments{
\item{bst}{boosting model from \code{lgb.train.cv}}

\item{df.test}{data.frame of test set}

\item{...}{predcit.lgb.Booster other params}
}
\description{
predict from xgboost model
}
