% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_xgboost.R
\name{xgb.save_model}
\alias{xgb.save_model}
\title{xgb.save_model}
\usage{
xgb.save_model(bst, model.dir, model.id = "", verbose = T)
}
\arguments{
\item{bst}{boosting model from \code{lgb.train.cv}}

\item{model.dir}{base path for saving model object and meta-data}

\item{model.id}{identifier for model}

\item{verbose}{whether display saving information, defaults at TRUE}
}
\description{
save Xgboost model from \code{xgb.train_with_cv}
}
