% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_xgboost.R
\name{xgb.train_with_cv}
\alias{xgb.train_with_cv}
\title{xgb.train_with_cv}
\usage{
xgb.train_with_cv(
  df.train,
  df.valid,
  fnames,
  label,
  fnames.cat,
  id = NULL,
  rules = NULL,
  params = NULL,
  cv.verbose = 0,
  train.verbose = 1,
  cv = FALSE,
  nfold = 5,
  score.fun = Metrics::auc,
  ...
)
}
\arguments{
\item{df.train}{data.frame for train set}

\item{df.valid}{data.frame for valid set}

\item{fnames}{all feature names}

\item{label}{label name}

\item{fnames.cat}{categorical feature names}

\item{id}{colnames of id, if not provide, default to use rowname}

\item{rules}{rules to index categorical features}

\item{params}{list of params for xgboost}

\item{cv.verbose}{whether verbose when cv}

\item{train.verbose}{whether verbose when train}

\item{cv}{whether perform a k-fold cross validation, default at FALSE}

\item{nfold}{number of folds of cross validation, if cv is TRUE}

\item{score.fun}{score functino for validation, defaults to auc}

\item{...}{other xgboost parameters}
}
\value{
trained xgboost model
}
\description{
Train Xgboost with Cross-valiation
}
\examples{
label <- 'label'
fnames <- c('Sex','Class','Age','Freq')
fnames.cat <- c('Sex','Age','Class')
df <- data.frame(Titanic)
df$label <- ifelse(df$Survived == "Yes", 1 ,0)
in.train <- runif(nrow(df)) < 0.8
df.train <- df[in.train, ]
df.valid <- df[!in.train, ]

params <- list(max_depth = 2, eta = 1,
               nthread = 2,objective = "binary:logistic",
               eval_metric = "auc")
bst <- xgb.train_with_cv(df.train, df.valid, fnames,
                       label, fnames.cat, params = params,
                       cv.verbose = 1, train.verbose = 1,
                       cv = TRUE, nfold = 2,
                       print_every_n = 5L,
                       early_stopping_rounds = 5L,
                       nrounds = 100L)

xgb.save_model(bst, 'saved_model','xgb_baseline')
bst_loaded <- xgb.load_model('saved_model','xgb_baseline')
preds <- xgb.predict(bst_loaded, df.valid)
}
